chore(pie-monorepo-utils): DSW-1234 move utility fns
